+++ prowler/lib/outputs/compliance/c5/models.py
@@ -55,3 +55,30 @@ class AzureC5Model(BaseModel):
     Muted: bool
     Framework: str
     Name: str
+
+
+class GCPC5Model(BaseModel):
+    """
+    GCPC5Model generates a finding's output in GCP C5 Compliance format.
+    """
+
+    Provider: str
+    Description: str
+    ProjectId: str
+    Location: str
+    AssessmentDate: str
+    Requirements_Id: str
+    Requirements_Description: str
+    Requirements_Attributes_Section: str
+    Requirements_Attributes_SubSection: str = None
+    Requirements_Attributes_Type: str = None
+    Requirements_Attributes_AboutCriteria: Optional[str] = None
+    Requirements_Attributes_ComplementaryCriteria: Optional[str] = None
+    Status: str
+    StatusExtended: str
+    ResourceId: str
+    ResourceName: str
+    CheckId: str
+    Muted: bool
+    Framework: str
+    Name: str